app/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
psycopg2-binary==2.9.11
redis==6.4.0
pymongo==4.15.3
boto3==1.40.53
prometheus-client==0.23.1
pydantic==2.12.2